refactor: move types file to types folder, create store folder for sprint context
